Start moving general adhoc stuff out of sceNetAdhoc
