+++ Core/HLE/sceNet.cpp
@@ -54,6 +54,9 @@
 #include "Core/HLE/sceNp2.h"
 #include "Core/HLE/sceNetInet.h"
 #include "Core/HLE/sceNetResolver.h"
+#include "Core/HLE/NetAdhocCommon.h"
+
+#include "Core/MIPS/MIPSCodeUtils.h" // for macros to implement __CreateHLELoop
 
 // These are all public. Should probably add accessors around these.
 bool g_netInited;
@@ -522,6 +525,19 @@ static void __ResetInitNetLib() {
 
 	memset(&netMallocStat, 0, sizeof(netMallocStat));
 	memset(&parameter, 0, sizeof(parameter));
+}
+
+u32_le __CreateHLELoop(u32_le *loopAddr, const char *sceFuncName, const char *hleFuncName, const char *tagName) {
+    if (loopAddr == NULL || sceFuncName == NULL || hleFuncName == NULL)
+        return 0;
+
+    loopAddr[0] = MIPS_MAKE_SYSCALL(sceFuncName, hleFuncName);
+    loopAddr[1] = MIPS_MAKE_B(-2);
+    loopAddr[2] = MIPS_MAKE_NOP();
+    u32 blockSize = sizeof(u32_le)*3;
+    u32_le threadHackAddress = kernelMemory.Alloc(blockSize, false, tagName); // blockSize will be rounded to 256 granularity
+    Memory::Memcpy(threadHackAddress, loopAddr, sizeof(u32_le) * 3); // This area will be cleared again after loading an old savestate :(
+    return threadHackAddress;
 }
 
 void __NetCallbackInit() {
